+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#184](https://github.com/deviceinsight/kafkactl/pull/184) Added option to show default configs when describing topics
 - [#183](https://github.com/deviceinsight/kafkactl/issues/183) Add command `delete records` to delete records from topic
 
+### Changed
+- [#180](https://github.com/deviceinsight/kafkactl/issues/180) Change default for replication-factor when creating topics
+
 ## 3.5.1 - 2023-11-10
 
 ## 3.5.0 - 2023-11-10